user_management-task/requirements.txt
# pytest==7.1.1
psycopg2==2.9.6
Django==4.2.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
